+++ modules/loader-utils/src/index.js
@@ -1,3 +1,5 @@
+/** @typedef {import('@loaders.gl/loader-utils').WorkerObject} WorkerObject */
+
 // GENERAL UTILS
 export {default as assert} from './lib/env-utils/assert';
 export {
@@ -16,6 +18,8 @@ export {validateLoaderVersion} from './lib/worker-loader-utils/validate-loader-v
 export {makeTransformIterator} from './lib/iterator-utils/make-transform-iterator';
 
 // WORKER UTILS
+export {processOnWorker} from './lib/worker-utils/process-on-worker';
+export {createWorker} from './lib/worker-utils/create-worker';
 export {getTransferList} from './lib/worker-utils/get-transfer-list';
 export {default as _WorkerFarm} from './lib/worker-utils/worker-farm';
 export {default as _WorkerPool} from './lib/worker-utils/worker-pool';
@@ -74,6 +78,20 @@ export {default as ChildProcessProxy} from './lib/process-utils/child-process-pr
 // MESH CATEGORY UTILS
 // Note: Should move to category specific module if code size increases
 export {getMeshSize as _getMeshSize, getMeshBoundingBox} from './categories/mesh/mesh-utils';
+
+// __VERSION__ is injected by babel-plugin-version-inline
+// @ts-ignore TS2304: Cannot find name '__VERSION__'.
+const VERSION = typeof __VERSION__ !== 'undefined' ? __VERSION__ : 'latest';
+
+export const NullWorker = {
+  id: 'null',
+  name: 'null',
+  module: 'loader-utils',
+  version: VERSION,
+  options: {
+    null: {}
+  }
+};
 
 // DEPRECATED IN 2.3
 export {getZeroOffsetArrayBuffer} from './lib/binary-utils/memory-copy-utils';