+++ modules/loader-utils/src/workers/null-worker.js
@@ -1,0 +1,6 @@
+import {createWorker} from '../lib/worker-utils/create-worker';
+
+createWorker(async ({data, options = {}}) => {
+  // @ts-ignore
+  return data;
+});